update_submodule_versions/requirements.txt
GitPython==3.1.29
ruamel.yaml==0.17.21
PyGithub==1.58.1
